+++ src/Identity/Logitar.Identity.Core/Sessions/ISessionRepository.cs
@@ -1,0 +1,9 @@
+﻿using Logitar.Identity.Domain.Sessions;
+
+namespace Logitar.Identity.Core.Sessions;
+
+public interface ISessionRepository
+{
+  Task SaveAsync(SessionAggregate session, CancellationToken cancellationToken = default);
+  Task SaveAsync(IEnumerable<SessionAggregate> sessions, CancellationToken cancellationToken = default);
+}