+++ src/Identity/Logitar.Identity.Core/Sessions/Payloads/CreateSessionPayload.cs
@@ -1,0 +1,8 @@
+﻿namespace Logitar.Identity.Core.Sessions.Payloads;
+
+public record CreateSessionPayload
+{
+  public string UserId { get; set; } = string.Empty;
+
+  public bool IsPersistent { get; set; }
+}