+++ src/Identity/tests/Logitar.Identity.IntegrationTests/IntegrationTestingBase.cs
@@ -45,6 +45,7 @@ public abstract class IntegrationTestingBase : IAsyncLifetime
     await EventContext.Database.ExecuteSqlRawAsync("DELETE FROM [dbo].[Events];");
 
     await IdentityContext.Database.MigrateAsync();
+    await IdentityContext.Database.ExecuteSqlRawAsync("DELETE FROM [dbo].[Sessions];");
     await IdentityContext.Database.ExecuteSqlRawAsync("DELETE FROM [dbo].[Users];");
     await IdentityContext.Database.ExecuteSqlRawAsync("DELETE FROM [dbo].[Tokenblacklist];");
   }